+++ src/ptm/DWR/DMS/PTM/SalmonSwimBehavior.java
@@ -6,7 +6,7 @@ public interface SalmonSwimBehavior extends SwimBehavior {
 	public float getSwimmingVelocity(int pId, int chanId);
 	public void setMeanSwimmingVelocity(int pId, int chanId);
 	public void setSwimmingTime(Particle p, int chanId);
-	public int getSwimmingTime(int pId, int chanId);
+	public long getSwimmingTime(int pId, int chanId);
 	public int getConfusionFactor(int chanId);
 	public void updatePosition(Particle p, float deltaT);
 	public void setXYZLocationInChannel(Particle p);
